requirements.txt
allure-pytest==2.12.0
pytest==7.0.1
pytest-xdist==3.0.2
requests==2.27.1
selenium==3.141.0
webdriver-manager==3.7.1
